mobile-messenger/requirements.txt
# mobile-messenger/requirements.txt
flet
python-dotenv
flet_desktop